Create separate initrd images for the smp and the non-smp kernels.
The size of the initrd.img went over the 32MB limit if modules for both
the smp and the non-smp kernels were to be included. My system refused
to boot with an initrd.img that big. To solve this, two separate initrd
files are created, one for the smp kernel and one for the non-smp
kernel. That will increase the total i486 size by about 30MB, but at
least it works.