Make it possible to have core-only or core/basic-only installations
If only a "core" directory is found in the installation medium, don't
offer and options about installation mode, just default to "CORE". If a
"full" directory is not found in the installation medium, offer the
option between "CORE" and "BASIC" only. Offer all option only if a
"full" directory is there.