Replace positioning variable name
Using $1 inside eval_gettext produces this warning:

warning: Although being used in a format string position, the msgid is
not a valid Shell format string. Reason: The string refers to a shell
variable whose value may be different inside shell functions.

So, we're creating a PART=$1 variable at the start of the function and
use that inside the eval_gettext strings.