Remove load_ramdisk, prompt_ramdisk boot options
They are not needed anymore.

prompt_ramdisk=0 in Slackware command line is superfluous as it is the
default as stated in /usr/src/linux/Documentation/blockdev/ramdisk. For
the records, this just says that as the kernel and the iniramfs are on
the same floppy disk (!) there is no need to give the user some time to
switch floppy disks before trying to load the ramdisk.

load_ramdisk=1 just says that a ramdisk should be loaded. I believe that
this was only needed when using floppy disks as well.