Read available locales from isolinux.cfg
Read the available locales from the 64-bit isolinux.cfg file. They
should be the same as in the 32-bit one anyway. For each locale, copy
all respective locale files from the glibc-i18n package inside the
initrd.

Downside is that if there is no menu option for a locale, that will not
be included in the initrd, therefore it cannot be used. I was thinking
that maybe we could include partial translations, don't show them at the
menu, but allow the user to manually set the LANG for them in the boot
options. However, that is not possible if the locales are determined
this way. Maybe not a big deal though, hardly any users would ever do
that.